Imported current state from SVN.
The current state includes functionality to output through a MessageDisplayer component into a TextWriter component, and then out to an ITextOutput, which will handle displaying it for the user (OutputToTMPro implements).
As it stands, TextWriter isn't even an ITextOutput yet, and the intention to make a new IMessageOutput interface to handle MessageDisplayer's output of a Message struct...may be a little overdone?

I dunno. MessageDisplayer (which should really be called MessageTokenizeComponent, if we're honest) could probably also be an ITextOutput--just have MessageQueue output to _any_ ITextOutput?
This would necessarily imply that less than Unity 2019.3 could not be allowed as options, which would suck.